Pass post ID into embed reversal endpoint, allow for filtering response
When posting an inserted embed code to the Ajax endpoint, we should also
send the current post ID, so that it is available as additional context
to the "reversed_embed" action.

This also introduces a filter on "shortcake_bakery_embed_reversal", so
that it's possible to filter the result of reversing an embed code. The
idea is that this can be used to implement a system of protected embeds,
or some other way of handling embed codes which don't match one of the
defined post elements.