Clean up and escape dimension attributes
All of these document shortcodes should have dimensions attributes, and
the ability to disable the "shortcake-bakery-responsive" attribute when
desired.

width and height are being escaped through `wp_kses_one_attr()`, which
is the best way I could find to late-escape these values. This is
probably terrible for performance, so it would be nice to do this in a
not-so-heavy way.