Conditionally show attributes based on "type" field
Adds a JS hook to change the fields displayed on the GoogleDocs
shortcode depending on the value of the "type" field.

It was necessary to add the shortcode names to the script localization,
because the shortcode tag names can be filtered or overridden, and the
hook that's triggered is keyed off of the tag name.