Add "height-adjust" to responive elements script
Because from iframes, like Instagram embeds, have a non-responsive
element (header + footer) as well as the responsive part (the image), it
would be helpful to be able to specify the two separately. This adds an
additional data-element (data-height-adjust) which can be specified on
the .shortcake-bakery-responsive element. If this is set, that portion
of the height will be considered "static" and will not be recalculated
on resize.