Add support for YouTube short urls (youtu.be/xxxxxxxxxx)
When trying to get a URL for a YouTube video, a logical place to look is
the "Share" link. The link provided if you click "Share", and then "Copy
URL", is a shortlink of the form `https://youtu.be/{embed_id}`.

This PR adds support for that url pattern to the Youtube post element.