Try to fix build by allowing latest version of PHPCS
The version of phpcodesniffer required by WPCS (^3.0.1) conflicted with
the pinned version of phpcs in composer.json. SInce phpcs is registered
as a dependency of wpcs, it makes sense to not pin it to a specific
version.

Also, PHP 5.3 isn't around anymore, and doesn't need to be included in
the Travis matrix.